requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.0
ddgs>=1.9.7
beautifulsoup4==4.12.2
python-dotenv==1.0.0
pydantic>=2.0.0,<3.0.0
requests==2.31.0
httpx
